requirements.txt
google-api-python-client
youtube-transcript-api
openai
python-dotenv